+++ core/templates/components/common-layout-directives/common-elements/attribution-guide.component.html
@@ -1,4 +1,4 @@
-<div class="oppia-attribution-guide">
+<div class="oppia-attribution-guide" *ngIf="!checkMobileView()">
   <div [ngStyle]="isLanguageRTL() ? {'left': '25px'} : {'right': '25px'}" class="oppia-cc-icon oppia-transition-200" *ngIf="!deviceUsedIsMobile && !iframed">
     <div class="oppia-neutral-element protractor-test-neutral-element"></div>
