+++ core/templates/components/common-layout-directives/common-elements/attribution-guide.component.spec.ts
@@ -27,6 +27,7 @@ import { AttributionService } from 'services/attribution.service';
 import { BrowserCheckerService } from 'domain/utilities/browser-checker.service';
 import { MockTranslatePipe } from 'tests/unit-test-utils';
 import { I18nLanguageCodeService } from 'services/i18n-language-code.service';
+import { WindowDimensionsService } from 'services/contextual/window-dimensions.service';
 
 class MockAttributionService {
   init() {
@@ -84,6 +85,7 @@ describe('Attribution Guide Component', function() {
   let component: AttributionGuideComponent;
   let fixture: ComponentFixture<AttributionGuideComponent>;
   let i18nLanguageCodeService: I18nLanguageCodeService;
+  let windowDimensionsService: WindowDimensionsService;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
@@ -95,7 +97,8 @@ describe('Attribution Guide Component', function() {
         { provide: AttributionService, useClass: MockAttributionService },
         { provide: BrowserCheckerService, useClass: MockBrowserCheckerService },
         { provide: UrlService, useClass: MockUrlService },
-        { provide: ContextService, useClass: MockContextService }
+        { provide: ContextService, useClass: MockContextService },
+        WindowDimensionsService
       ],
       schemas: [NO_ERRORS_SCHEMA]
     }).compileComponents();
@@ -104,11 +107,22 @@ describe('Attribution Guide Component', function() {
   beforeEach(() => {
     fixture = TestBed.createComponent(AttributionGuideComponent);
     i18nLanguageCodeService = TestBed.inject(I18nLanguageCodeService);
+    windowDimensionsService = TestBed.inject(WindowDimensionsService);
 
     component = fixture.componentInstance;
     spyOn(i18nLanguageCodeService, 'isCurrentLanguageRTL').and.returnValue(
       true);
   });
+
+  it('should return true if the window size is less than equal to 1024px',
+    function() {
+      let widthSpy = spyOn(windowDimensionsService, 'getWidth');
+      widthSpy.and.returnValue(400);
+      expect(component.checkMobileView()).toBe(true);
+
+      widthSpy.and.returnValue(1025);
+      expect(component.checkMobileView()).toBe(false);
+    });
 
   it('should initialize component properties correctly', () => {
     expect(component.deviceUsedIsMobile).toBeFalse();
