+++ core/templates/components/common-layout-directives/common-elements/attribution-guide.component.ts
@@ -25,6 +25,7 @@ import { AttributionService } from 'services/attribution.service';
 import { ContextService } from 'services/context.service';
 import { UrlService } from 'services/contextual/url.service';
 import { I18nLanguageCodeService } from 'services/i18n-language-code.service';
+import { WindowDimensionsService } from 'services/contextual/window-dimensions.service';
 
 @Component({
   selector: 'attribution-guide',
@@ -42,7 +43,8 @@ export class AttributionGuideComponent implements OnInit {
     private browserCheckerService: BrowserCheckerService,
     private contextService: ContextService,
     private i18nLanguageCodeService: I18nLanguageCodeService,
-    private urlService: UrlService
+    private urlService: UrlService,
+    private windowDimensionsService: WindowDimensionsService
   ) {}
 
   ngOnInit(): void {
@@ -59,6 +61,10 @@ export class AttributionGuideComponent implements OnInit {
     if (this.generateAttibutionIsAllowed) {
       this.attributionService.init();
     }
+  }
+
+  checkMobileView(): boolean {
+    return (this.windowDimensionsService.getWidth() <= 1024);
   }
 
   getAttributionModalStatus(): boolean {
