+++ core/templates/components/question-directives/question-player/question-player.component.html
@@ -15,7 +15,7 @@
     </div>
     <div class="oppia-question-player-body">
       <conversation-skin question-player-config="$ctrl.questionPlayerConfig"></conversation-skin>
-      <attribution-guide ng-if="!$ctrl.checkMobileView()"></attribution-guide>
+      <attribution-guide></attribution-guide>
     </div>
   </div>
   <!-- End of Question Player -->
