+++ core/templates/pages/collection-player-page/collection-player-page.directive.html
@@ -249,7 +249,7 @@
     </div>
   </div>
 </div>
-<attribution-guide ng-if="!$ctrl.checkMobileView()"></attribution-guide>
+<attribution-guide></attribution-guide>
 
 <style>
   @media screen and (max-width: 942px) {
