+++ core/templates/pages/exploration-editor-page/editor-tab/exploration-editor-tab.component.html
@@ -47,7 +47,7 @@
       <unresolved-answers-overview></unresolved-answers-overview>
     </div>
   </div>
-  <attribution-guide ng-if="!$ctrl.checkMobileView()"></attribution-guide>
+  <attribution-guide></attribution-guide>
 </div>
 
 <style>
