+++ core/templates/pages/exploration-editor-page/translation-tab/translation-tab.component.html
@@ -15,6 +15,6 @@
         <state-translation-status-graph is-translation-tab-busy="isTranslationTabBusy"></state-translation-status-graph>
       </div>
     </div>
-    <attribution-guide ng-if="!checkMobileView()"></attribution-guide>
+    <attribution-guide></attribution-guide>
   </div>
 </div>
