+++ core/templates/pages/exploration-player-page/exploration-player-page.component.html
@@ -1,6 +1,6 @@
 <div>
   <background-banner></background-banner>
   <conversation-skin></conversation-skin>
-  <attribution-guide ng-if="!$ctrl.checkMobileView()"></attribution-guide>
+  <attribution-guide></attribution-guide>
   <oppia-on-screen-keyboard></oppia-on-screen-keyboard>
 </div>
