+++ core/templates/pages/exploration-player-page/exploration-player-page.component.spec.ts
@@ -31,8 +31,6 @@ require('pages/exploration-player-page/exploration-player-page.component.ts');
 describe('Exploration player page', function() {
   var ctrl = null;
   var $q = null;
-  var $window = null;
-  var $rootScope = null;
   var $scope = null;
   var ContextService = null;
   var PageTitleService = null;
@@ -54,8 +52,7 @@ describe('Exploration player page', function() {
 
   beforeEach(angular.mock.inject(function($injector, $componentController) {
     $q = $injector.get('$q');
-    $rootScope = $injector.get('$rootScope');
-    $window = $injector.get('$window');
+    var $rootScope = $injector.get('$rootScope');
     ContextService = $injector.get('ContextService');
     PageTitleService = $injector.get('PageTitleService');
     ReadOnlyExplorationBackendApiService = $injector.get(
@@ -64,10 +61,6 @@ describe('Exploration player page', function() {
     $scope = $rootScope.$new();
     ctrl = $componentController('explorationPlayerPage', {
       $scope: $scope
-    });
-    Object.defineProperty($window, 'innerWidth', {
-      get: () => undefined,
-      set: () => {}
     });
   }));
 
@@ -113,18 +106,4 @@ describe('Exploration player page', function() {
     expect(elementDescriptionProperty.attr('content')).toBe(
       exploration.objective);
   });
-
-  it('should return true if the window size is less than equal to 1024px',
-    () => {
-      var innerWidthSpy = spyOnProperty($window, 'innerWidth');
-      innerWidthSpy.and.callFake(() => 480);
-      $rootScope.$apply();
-      angular.element($window).triggerHandler('resize');
-      expect(ctrl.checkMobileView()).toBe(true);
-
-      innerWidthSpy.and.callFake(() => 1025);
-      $rootScope.$apply();
-      angular.element($window).triggerHandler('resize');
-      expect(ctrl.checkMobileView()).toBe(false);
-    });
 });
