+++ core/templates/pages/exploration-player-page/exploration-player-page.component.ts
@@ -37,10 +37,10 @@ require('services/keyboard-shortcut.service.ts');
 angular.module('oppia').component('explorationPlayerPage', {
   template: require('./exploration-player-page.component.html'),
   controller: [
-    '$rootScope', '$window', 'ContextService', 'KeyboardShortcutService',
+    '$rootScope', 'ContextService', 'KeyboardShortcutService',
     'PageTitleService', 'ReadOnlyExplorationBackendApiService',
     function(
-        $rootScope, $window, ContextService, KeyboardShortcutService,
+        $rootScope, ContextService, KeyboardShortcutService,
         PageTitleService, ReadOnlyExplorationBackendApiService) {
       var ctrl = this;
       ctrl.$onInit = function() {
@@ -61,10 +61,6 @@ angular.module('oppia').component('explorationPlayerPage', {
             $rootScope.$applyAsync();
           });
         KeyboardShortcutService.bindExplorationPlayerShortcuts();
-      };
-
-      ctrl.checkMobileView = function() {
-        return ($window.innerWidth <= 1024);
       };
     }
   ]
